refactor: title컴포넌트 생성,Radio/SeriesEditor컴포넌트에 적용
