+++ src/components/index.jsx
@@ -20,9 +20,10 @@ export { default as PageSectionTitle } from './domain/PageSectionTitle';
 export { default as ArticleList } from './domain/ArticleList';
 export { default as CommentList } from './domain/CommentList';
 export { default as Radio } from './commons/Radio/ButtonRadio';
-export { default as SeriesEditor } from './domain/SeriesEditor';
+export { default as SeriesEditor } from './domain/writeSeries/SeriesEditor';
 export { default as ConfirmCancleButtons } from './domain/ConfirmCancleButtons';
 export { default as ArticleEditor } from './domain/EditArticleForm/ArticleEditor';
 export { default as CardSlider } from './domain/CardSlider';
 export { default as EditArticleForm } from './domain/EditArticleForm';
 export { default as ImageUpload } from './domain/ImageUpload';
+export { default as Title } from './commons/Title';
